cli/pyproject.toml
[tool.poetry]
name = "gfuzz"
version = "1.0.0"
description = ""
authors = ["hgarrereyn"]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.4.1"
beautifulsoup4 = "^4.9.3"
tqdm = "^4.62.0"
pyparsing = "^2.4.7"
dataclasses = {version = "^0.8", python = ">=3.6, <3.7"}

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
gfuzz = "gfuzz.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
